// servidor-local/src/utils/types.d.ts
export interface EmpresaDBType {
    id: string;
    designacao: string;
    descricao: string;
    localizacao: string;
    nif: string;
    icone: string;
    id_utilizador: string;
    enabled: boolean;
    created_at: string;
    updated_at: string;
}